pretraining/bert/requirements.txt
# for pretraining bert
bunkai~=1.3.0
pytextspan>=0.5.4
tensorflow_datasets>=4.3.0
tokenizers>=0.10.3
progressbar~=2.5
tensorflow>=2.5.0
